# Conflicts:
#	package.json

+++ src/drugPlugin/viewer.js
@@ -1,3 +1,4 @@
+/* global window */
 "use strict";
 
 var Widget = require('./../ui/widget').Widget,
@@ -366,11 +367,13 @@ var Viewer = exports.Viewer = Widget.extend({
     //
     // Returns nothing.
     _onDeleteClick: function (event) {
-        var item = $(event.target)
-            .parents('.annotator-annotation')
-            .data('annotation');
-        this.hide();
-        this.options.onDelete(item);
+        if (window.confirm(_t('Delete this annotation?'))) {
+            var item = $(event.target)
+                .parents('.annotator-annotation')
+                .data('annotation');
+            this.hide();
+            this.options.onDelete(item);
+        }
     },
     // Event callback: called when the cancel button is clicked.
     //
